chore(new-demo): now it's `@ckb-ccc/app`
