+++ packages/app/README.md
@@ -1,0 +1,21 @@
+# CCC App
+
+The interactive CCC application and reference implementation for connecting
+wallets, inspecting accounts, and running common CKB workflows with CCC.
+
+## Development
+
+From the repository root:
+
+```bash
+pnpm --dir packages/app dev
+```
+
+Then open <http://localhost:3000>.
+
+## Package
+
+- Package name: `@ckb-ccc/app`
+- Source: `packages/app`
+- The legacy implementation remains available in `packages/demo` while the
+  application is being validated.
