+++ packages/app/next.config.mjs
