+++ packages/app/src/app/activity-console.tsx
