+++ packages/app/src/app/copy-text.ts
