+++ packages/app/src/app/copyable-readout-value.tsx
