+++ packages/app/src/app/modules/hash-module.tsx
